customizer api, crud, document, and fixes

+++ altrpnjs/app/Generators/ListenerGenerator.ts
@@ -32,6 +32,8 @@ export default class ListenerGenerator extends BaseGenerator {
 
     this.model = model;
 
+    console.log(model)
+
     const dir = ListenerGenerator.directory + "altrp_models" + "." + this.model.name + "." + type;
 
     if (!fs.existsSync(dir)){
@@ -40,6 +42,7 @@ export default class ListenerGenerator extends BaseGenerator {
 
     const listener = type
 
+    console.log(listener + ListenerGenerator.ext)
     await this.addFile( listener + ListenerGenerator.ext)
       .destinationDir(dir)
       .stub(ListenerGenerator.template)
@@ -63,7 +66,9 @@ export default class ListenerGenerator extends BaseGenerator {
 
     const listener = type
 
-    fs.unlinkSync(dir + "\\" + listener + ListenerGenerator.ext)
+    if(fs.existsSync(dir + "\\" + listener + ListenerGenerator.ext)) {
+      fs.unlinkSync(dir + "\\" + listener + ListenerGenerator.ext)
+    }
   }
 
   public async runAll(model) {
